compound finance section fixes
